refactor moveHead to be submodule of snake/snake
